Adding ComplexKeys and more tests for the Query class.
The ComplexKey objects make it possible to pass in objects and values
for those Query arguments, that need JSON as the input. This greatly
simplifies working with them.

The "update_seq" argument is removed, according to MB-6584.

Also, a new unit-test class for the Query objects is implemented
to verify the provided features.

Change-Id: I0e4a7b8816e14fbea0da9b60274e7270bb299631